+++ app/src/main/java/com/example/musicdiscovery/network/DeezerApiService.kt
@@ -12,9 +12,9 @@ interface DeezerApiService {
     @GET("search/artist")
     suspend fun getArtists(@Query("q") artistName: String): DeezerResponseList<Artist>
     @GET("artist/{artistId}/top")
-    suspend fun getArtistTracks(@Path("artistId") artistId: Int, @Query("limit") limit: Int, @Query("index") index: Int): DeezerResponseList<Track>
+    suspend fun getArtistTracks(@Path("artistId") artistId: Long, @Query("limit") limit: Int, @Query("index") index: Int): DeezerResponseList<Track>
     @GET("track/{trackId}")
     suspend fun getTrackDetails(@Path("trackId") trackId: Long): TrackDetails
     @GET("artist/{artistId}")
-    suspend fun getArtistDetails(@Path("artistId") artistId: Int): Artist
+    suspend fun getArtistDetails(@Path("artistId") artistId: Long): Artist
 }
